deprecated AbstractActions. moved methods to the generators
